fixed functions
