+++ guide/template.py
@@ -1,4 +1,6 @@
-from functions import *
+from st2334_helper import general as gn
+from st2334_helper import confidence_intervals as ci
+from st2334_helper import hypotheses_tests as ht
 
 # Questions below
 
@@ -6,7 +8,7 @@ def q1():
     mu = 0
     sd = 1
     ans = 0
-    disp(ans)
+    gn.disp(ans)
 
 # Generate solutions below
