+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = ST2334_Helper
-version = 0.1.8
+version = 0.1.14
 author = Prittam Ravi & Kevin Chang Jon Kit
 author_email = kevinchangjk@gmail.com
 description = A tool for NUS module ST2334
@@ -20,9 +20,9 @@ classifiers =
 package_dir =
   = src
 packages = find:
-python_requires = >=3.6
-install_requires = scipy
-
+python_requires = >=3.9
+install_requires =
+  scipy>=1.8.0
 
 [options.packages.find]
 where = src
