+++ src/ST2334_Helper.egg-info/requires.txt
@@ -1,1 +1,1 @@
-scipy
+scipy>=1.8.0
